# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Interacción con Base de Datos FireBird",
        "description": "API para la gestión",
        "version": "1.0"
    },
    "paths": {
        "/api/articulos": {
            "get": {
                "tags": [
                    "Articulo"
                ],
                "summary": "Obtener todos los artículos",
                "operationId": "021362b92b70d8c24f566bb0db7178d5",
                "responses": {
                    "200": {
                        "description": "Lista de artículos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Articulo"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Articulo"
                ],
                "summary": "Crear un nuevo artículo",
                "operationId": "1576d3467aeea24559ab2a70e9a963d9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Articulo"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Artículo creado exitosamente"
                    }
                }
            }
        },
        "/api/articulos/{id}": {
            "get": {
                "tags": [
                    "Articulo"
                ],
                "summary": "Obtener un artículo por ID",
                "operationId": "53ebb83e59d006abd1bd488176d98899",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del artículo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Información del artículo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Articulo"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Articulo"
                ],
                "summary": "Actualizar un artículo",
                "operationId": "8b012fffca1a2f25e37727c1097c0471",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del artículo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Articulo"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Artículo actualizado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Articulo"
                ],
                "summary": "Eliminar un artículo",
                "operationId": "de4809e8ad8eea751584f66ed2aab554",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del artículo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Artículo eliminado"
                    }
                }
            }
        },
...
        "/api/articulos/restore/{id}": {
            "put": {
                "tags": [
                    "Articulo"
                ],
                "summary": "Restaurar un artículo eliminado",
                "operationId": "7bbccb6309168f9f5e5d456b192bb12c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del artículo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Artículo restaurado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Articulo": {
                "title": "Articulo",
                "description": "Modelo de Artículo",
                "required": [
                    "ID_INVENTARIOS",
                    "NOMBRE",
                    "MARCA",
                    "DESCRIPCION",
                    "FECHACADUCIDAD",
                    "UNIDAD",
                    "CANTIDAD"
                ],
                "properties": {
                    "ID_INVENTARIOS": {
                        "description": "ID del inventario asociado",
                        "type": "integer"
                    },
                    "NOMBRE": {
                        "description": "Nombre del artículo",
                        "type": "string"
                    },
                    "MARCA": {
                        "description": "Marca del artículo",
                        "type": "string"
                    },
                    "DESCRIPCION": {
                        "description": "Descripción del artículo",
                        "type": "string"
                    },
                    "FECHACADUCIDAD": {
                        "description": "Fecha de caducidad del artículo",
                        "type": "string",
                        "format": "date"
                    },
                    "UNIDAD": {
                        "description": "Unidad de medida del artículo",
                        "type": "string"
                    },
                    "CANTIDAD": {
                        "description": "Cantidad disponible del artículo",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}